Fix sidekiq jobs errors

+++ app/jobs/delete_resource_files_job.rb
@@ -16,10 +16,14 @@ class DeleteResourceFilesJob < BaseJob
     if resource.download_uri&.include?('ark%3A%2F')
       m = /ark.*/.match(resource.download_uri)
       base_path = CGI.unescape(m.to_s)
-      merritt_version = resource.stash_version.merritt_version
-      perm_bucket.delete_dir(s3_key: "#{base_path}|#{merritt_version}|producer")
-      perm_bucket.delete_dir(s3_key: "#{base_path}|#{merritt_version}|system")
-      perm_bucket.delete_file(s3_key: "#{base_path}|manifest")
+      merritt_version = resource.stash_version&.merritt_version
+      if(merritt_version)
+        perm_bucket.delete_dir(s3_key: "#{base_path}|#{merritt_version}|producer")
+        perm_bucket.delete_dir(s3_key: "#{base_path}|#{merritt_version}|system")
+        perm_bucket.delete_file(s3_key: "#{base_path}|manifest")
+      else
+        puts "Resource delete - No merritt version for #{resource.id}"
+      end
     else
       perm_bucket.delete_dir(s3_key: "v3/#{s3_dir}")
     end
